rename `gpt_system` to `llm_system`
